+++ app/src/main/java/com/marat/hvatit/playlistmaker2/MainActivity.kt
@@ -4,61 +4,10 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
-import com.marat.hvatit.playlistmaker2.models.Track
 
 class MainActivity : AppCompatActivity() {
 
-    companion object {
-        val data: ArrayList<Track> = initialize()
 
-        private fun initialize(): ArrayList<Track> {
-            val songList: ArrayList<Track> = arrayListOf()
-            songList.add(
-                Track(
-                    "Smells Like Teen Spirit",
-                    "Nirvana",
-                    "5:01",
-                    "https://is5-ssl.mzstatic.com/image/thumb/Music115/v4/7b/58/c2/7b58c21a-2b51-2bb2-e59a-9bb9b96ad8c3/00602567924166.rgb.jpg/100x100bb.jpg"
-                )
-            )
-            songList.add(
-                Track(
-                    "Billie Jean",
-                    "Michael Jackson",
-                    "4:35",
-                    "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/3d/9d/38/3d9d3811-71f0-3a0e-1ada-3004e56ff852/827969428726.jpg/100x100bb.jpg"
-                )
-            )
-
-            songList.add(
-                Track(
-                    "Stayin' Alive",
-                    "Bee Gees",
-                    "4:10",
-                    "https://is4-ssl.mzstatic.com/image/thumb/Music115/v4/1f/80/1f/1f801fc1-8c0f-ea3e-d3e5-387c6619619e/16UMGIM86640.rgb.jpg/100x100bb.jpg"
-                )
-            )
-
-            songList.add(
-                Track(
-                    "Whole Lotta Love",
-                    "Led Zeppelin",
-                    "5:33",
-                    "https://is2-ssl.mzstatic.com/image/thumb/Music62/v4/7e/17/e3/7e17e33f-2efa-2a36-e916-7f808576cf6b/mzm.fyigqcbs.jpg/100x100bb.jpg"
-                )
-            )
-
-            songList.add(
-                Track(
-                    "Sweet Child O'Mine",
-                    "Guns N' Roses",
-                    "5:03",
-                    "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/a0/4d/c4/a04dc484-03cc-02aa-fa82-5334fcb4bc16/18UMGIM24878.rgb.jpg/100x100bb.jpg"
-                )
-            )
-            return songList
-        }
-    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -66,12 +15,9 @@ class MainActivity : AppCompatActivity() {
 
         val buttonSearch = findViewById<Button>(R.id.button_bigOne)
         buttonSearch.setOnClickListener {
-            SearchActivity.getIntent(this@MainActivity, this.getString(R.string.android))
-                .apply { putExtra(
-                    "data",
-                    Companion.data
-                )
-                    startActivity(this) }
+            SearchActivity.getIntent(this@MainActivity, this.getString(R.string.android)).apply {
+                    startActivity(this)
+                }
         }
         //................................................................................
         val buttonMedialib = findViewById<Button>(R.id.button_bigTwo)
