+++ app/src/main/java/com/marat/hvatit/playlistmaker2/SearchActivity.kt
@@ -5,7 +5,6 @@ import android.content.Intent
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
-import android.util.Log
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 import android.widget.EditText
@@ -15,6 +14,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.marat.hvatit.playlistmaker2.adapters.TrackListAdapter
 import com.marat.hvatit.playlistmaker2.models.Track
+import com.bumptech.glide.Glide
 
 const val EDITTEXT_TEXT = "EDITTEXT_TEXT"
 private const val TAG = "SearchActivity"
@@ -24,6 +24,56 @@ class SearchActivity : AppCompatActivity() {
     private var saveEditText: String = "error"
 
     companion object {
+
+        val data: ArrayList<Track> = initialize()
+
+        private fun initialize(): ArrayList<Track> {
+            val songList: ArrayList<Track> = arrayListOf()
+            songList.add(
+                Track(
+                    "Smells Like Teen Spirit",
+                    "Nirvana",
+                    "5:01",
+                    "https://is5-ssl.mzstatic.com/image/thumb/Music115/v4/7b/58/c2/7b58c21a-2b51-2bb2-e59a-9bb9b96ad8c3/00602567924166.rgb.jpg/100x100bb.jpg"
+                )
+            )
+            songList.add(
+                Track(
+                    "Billie Jean",
+                    "Michael Jackson",
+                    "4:35",
+                    "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/3d/9d/38/3d9d3811-71f0-3a0e-1ada-3004e56ff852/827969428726.jpg/100x100bb.jpg"
+                )
+            )
+
+            songList.add(
+                Track(
+                    "Stayin' Alive",
+                    "Bee Gees",
+                    "4:10",
+                    "https://is4-ssl.mzstatic.com/image/thumb/Music115/v4/1f/80/1f/1f801fc1-8c0f-ea3e-d3e5-387c6619619e/16UMGIM86640.rgb.jpg/100x100bb.jpg"
+                )
+            )
+
+            songList.add(
+                Track(
+                    "Whole Lotta Love",
+                    "Led Zeppelin",
+                    "5:33",
+                    "https://is2-ssl.mzstatic.com/image/thumb/Music62/v4/7e/17/e3/7e17e33f-2efa-2a36-e916-7f808576cf6b/mzm.fyigqcbs.jpg/100x100bb.jpg"
+                )
+            )
+
+            songList.add(
+                Track(
+                    "Sweet Child O'Mine",
+                    "Guns N' Roses",
+                    "5:03",
+                    "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/a0/4d/c4/a04dc484-03cc-02aa-fa82-5334fcb4bc16/18UMGIM24878.rgb.jpg/100x100bb.jpg"
+                )
+            )
+            return songList
+        }
 
         fun getIntent(context: Context, message: String): Intent {
             return Intent(context, SearchActivity::class.java).apply {
@@ -73,13 +123,9 @@ class SearchActivity : AppCompatActivity() {
                 0
             )
         }
-        val arraylist: ArrayList<Track>? =
-            Companion.getIntent(this,this.getString(R.string.android)).getParcelableExtra("data") as ArrayList<Track>?
-        //val intent: Intent = getIntent(this@SearchActivity, "data")
         val recyclerSongList = findViewById<RecyclerView>(R.id.songlist)
         recyclerSongList.layoutManager = LinearLayoutManager(this)
-        recyclerSongList.adapter = arraylist?.let { TrackListAdapter(tracklist = it) }
-
+        recyclerSongList.adapter = TrackListAdapter(data)
 
     }
